LPS-55168 moves java
