+++ modules/apps/site/site-memberships-portlet/src/com/liferay/site/memberships/web/portlet/SiteMembershipsControlPanelEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.sitesadmin;
+package com.liferay.site.memberships.web.portlet;
 
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.Portlet;
